remove barrel files
