+++ client/src/features/projectsV2/LazyProjectV2New.tsx
@@ -1,5 +1,5 @@
 /*!
- * Copyright 2023 - Swiss Data Science Center (SDSC)
+ * Copyright 2024 - Swiss Data Science Center (SDSC)
  * A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
  * Eidgenössische Technische Hochschule Zürich (ETHZ).
  *
@@ -13,17 +13,17 @@
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
- * limitations under the License.
+ * limitations under the License
  */
+import { Suspense, lazy } from "react";
+import PageLoader from "../../components/PageLoader";
 
-export {
-  ProjectDescriptionFormField,
-  ProjectNameFormField,
-  ProjectSlugFormField,
-  ProjectVisibilityFormField,
-} from "./simpleFormFields";
+const ProjectV2New = lazy(() => import("./new/ProjectV2New"));
 
-export { AddProjectMemberModal } from "./projectMemberFields";
-import ProjectRepositoryFormField from "./ProjectRepositoryFormField";
-
-export { ProjectRepositoryFormField };
+export default function LazyProjectV2New() {
+  return (
+    <Suspense fallback={<PageLoader />}>
+      <ProjectV2New />
+    </Suspense>
+  );
+}
