--- client/src/features/projectsV2/index.ts
@@ -1,7 +1,0 @@
-export { projectV2Api } from "./api/";
-export { projectV2NewSlice } from "./new/projectV2New.slice";
-
-import ProjectV2List from "./list/ProjectV2List";
-import ProjectV2New from "./new/ProjectV2New";
-import ProjectV2Show from "./show/ProjectV2Show";
-export { ProjectV2List, ProjectV2New, ProjectV2Show };
