+++ client/src/features/projectsV2/show/ProjectV2Show.tsx
@@ -31,8 +31,11 @@ import { Loader } from "../../../components/Loader";
 import { TimeCaption } from "../../../components/TimeCaption";
 import { Url } from "../../../utils/helpers/url";
 
-import { isErrorResponse, useGetProjectsByProjectIdQuery } from "../api";
-import type { Project } from "../api";
+import {
+  isErrorResponse,
+  useGetProjectsByProjectIdQuery,
+} from "../api/projectV2.enhanced-api";
+import type { Project } from "../api/projectV2.api";
 import WipBadge from "../shared/WipBadge";
 
 import {
